pr updates

+++ xrdsst/main.py
@@ -17,23 +17,26 @@ META['output.json']['overridable'] = True
 META['output.tabulate']['overridable'] = True
 
 
-def revert_api_key(app):
-    config_file = app.Meta.handlers[0].config_file
-    api_key_id = app.Meta.handlers[0].api_key_id
-    if not os.path.exists(config_file):
-        config_file = os.path.join("..", config_file)
-    with open(config_file, "r") as yml_file:
-        config = yaml.load(yml_file, Loader=yaml.FullLoader)
-    for security_server in config["security-server"]:
-        logging.info('Revoking API key for security server ' + security_server['name'])
-        print('Revoking API key for security server ' + security_server['name'])
-        curl_cmd = "curl -X DELETE -u " + config["api-key"][0]["credentials"] + " --silent " + \
-                   config["api-key"][0]["url"] + "/" + str(api_key_id[security_server['name']]) + " -k"
-        cmd = "ssh -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no -o LogLevel=ERROR -i \"" + \
-            config["api-key"][0]["key"] + "\" root@" + security_server["name"] + " \"" + curl_cmd + "\""
-        process = subprocess.run(cmd, shell=True, check=False, capture_output=True)
-        logging.info('API key for security server ' + security_server['name'] + ' revoked successfully')
-        print('API key for security server ' + security_server['name'] + ' revoked successfully')
+def revoke_api_key(app):
+    if app.argv:
+        config_file = app.Meta.handlers[0].config_file
+        api_key_id = app.Meta.handlers[0].api_key_id
+        api_key_default = app.Meta.handlers[0].api_key_default
+        if not os.path.exists(config_file):
+            config_file = os.path.join("..", config_file)
+        with open(config_file, "r") as yml_file:
+            config = yaml.load(yml_file, Loader=yaml.FullLoader)
+        for security_server in config["security-server"]:
+            if security_server["api_key"] == api_key_default:
+                logging.info('Revoking API key for security server ' + security_server['name'])
+                print('Revoking API key for security server ' + security_server['name'])
+                curl_cmd = "curl -X DELETE -u " + config["api-key"][0]["credentials"] + " --silent " + \
+                           config["api-key"][0]["url"] + "/" + str(api_key_id[security_server['name']]) + " -k"
+                cmd = "ssh -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no -o LogLevel=ERROR -i \"" + \
+                    config["api-key"][0]["key"] + "\" root@" + security_server["name"] + " \"" + curl_cmd + "\""
+                process = subprocess.run(cmd, shell=True, check=False, capture_output=True)
+                logging.info('API key for security server ' + security_server['name'] + ' revoked successfully')
+                print('API key for security server ' + security_server['name'] + ' revoked successfully')
 
 
 class XRDSST(App):
@@ -68,7 +71,7 @@ class XRDSSTTest(TestApp, XRDSST):
 
 def main():
     with XRDSST() as app:
-        app.hook.register('pre_close', revert_api_key)
+        app.hook.register('pre_close', revoke_api_key)
         try:
             app.run()
         except AssertionError as err:
